Splitting the fucntions between cost matrix and actual SA.
