+++ src/Makefile.am
@@ -2,5 +2,5 @@ INCLUDES		=  -I$(top_builddir) -I$(top_srcdir)
 
 
 pkglib_LIBRARIES 	= librgraph.a
-librgraph_a_SOURCES 	= datastruct.c tools.c graph.c models.c modules.c bipartite.c missing.c matrix.c layout.c recommend.c conflict.c genetic.c louvain.c sannealing.c
-pkginclude_HEADERS 	= datastruct.h tools.h graph.h models.h modules.h bipartite.h missing.h matrix.h layout.h recommend.h conflict.h genetic.h louvain.h sannealing.h
+librgraph_a_SOURCES 	= datastruct.c tools.c graph.c models.c modules.c bipartite.c missing.c matrix.c layout.c recommend.c conflict.c genetic.c louvain.c sannealing.c costmatrix.c
+pkginclude_HEADERS 	= datastruct.h tools.h graph.h models.h modules.h bipartite.h missing.h matrix.h layout.h recommend.h conflict.h genetic.h louvain.h sannealing.h costmatrix.h
