+++ src/costmatrix.h
@@ -1,0 +1,7 @@
+#include "graph.h"
+#include "bipartite.h"
+
+double **CostMatrix(struct node_gra *net);
+double **CostMatrixWeighted(struct node_gra *net);
+double **CostMatrixBipart(struct binet *binet);
+double **CostMatrixBipartWeighted(struct binet *binet);
