+++ src/sannealing.h
@@ -11,22 +11,54 @@ struct partition{
   unsigned int nempty;
 };
 
-double **CostMatrix(struct node_gra *net);
-double **CostMatrixWeighted(struct node_gra *net);
-double **CostMatrixBipart(struct binet *binet);
-double **CostMatrixBipartWeighted(struct binet *binet);
 
-void iterationNumber(int nnod,
-					 double fac,
-					 int *individual_movements,
-					 int *collective_movements);
+void iterationNumber(int nnod, double fac,
+					 int *individual_movements, int *collective_movements);
 struct group *convertPartitionToGroup(struct partition *part, struct node_gra *net);
-unsigned int move(unsigned int val,
+struct partition *InitializePartition(unsigned int N);
+struct partition *CopySimplePartition(struct partition *part);
+void FreeSimplePartition(struct partition *part);
+
+
+/////////////////////////////
+unsigned int
+move(unsigned int val,
 				  unsigned int old, unsigned int new,
 				  struct partition *part);
-void merge(unsigned int g1,
+void
+merge(unsigned int g1,
 		   unsigned int g2,
 		   struct partition *part);
+
+double
+dEMergeGroups(unsigned int group1,
+			  unsigned int group2,
+			  struct partition *part,
+			  double **cmat);
+double
+dEAddNodeToGroup(unsigned int nodeid,
+				 unsigned int groupid,
+				 struct partition *part,
+				 double **cmat);
+
+/////////////////////////////////////
+
+
+struct group *
+GeneralSA(double **cmat, unsigned int N,
+		  struct node_gra *net,
+		  double fac,
+		  double Ti, double Tf, double Ts,
+		  gsl_rng *gen);
+
+void
+GeneralSAGroupSplit(unsigned int target, unsigned int empty,
+					double Ti, double Tf, double Ts, double fac,
+					double cluster_prob, unsigned int nochange_limit,
+					double **cmat, struct partition *part,
+					struct node_gra *net, gsl_rng *gen);
+
+/////////////////////////////////////////////////////////////:
 struct group *
 SACommunityIdent(struct node_gra *net,
 				   double Ti, double Tf, double Ts,
@@ -65,23 +97,3 @@ SACommunityIdentBipartWeighted(struct binet *binet,
 
 
 
-struct partition * InitializePartition(unsigned int N);
-struct partition *CopySimplePartition(struct partition *part);
-void FreeSimplePartition(struct partition *part);
-
-struct group *
-GeneralSA(double **cmat, unsigned int N,
-		  struct node_gra *net,
-		  double fac,
-		  double Ti, double Tf, double Ts,
-		  gsl_rng *gen);
-double
-dEMergeGroups(unsigned int group1,
-			  unsigned int group2,
-			  struct partition *part,
-			  double **cmat);
-double
-dEAddNodeToGroup(unsigned int nodeid,
-				 unsigned int groupid,
-				 struct partition *part,
-				 double **cmat);
